+++ R/remove_brackets.R
@@ -15,7 +15,7 @@
 #' )
 #'
 #' remove_brackets(
-#'   c("r-project.org <https://r-project.org>", "python.org <https://python.org>"),
+#'   c("R <https://r-project.org>", "Python <https://python.org>"),
 #'   bracket = "<"
 #' )
 #'
@@ -26,15 +26,18 @@
 remove_brackets <- function(string, bracket = c("(", "<", "[")) {
   bracket <- match.arg(bracket)
 
-  closing_bracket <- switch(bracket,
+  closing_bracket <- switch(
+    bracket,
     "<" = ">",
     "(" = ")",
     "[" = "]"
   )
 
-  string |>
-    stringi::stri_replace_all_regex(
-      paste0("\\", bracket, "[^", closing_bracket, "]*\\", closing_bracket),
-      ""
-    )
+  unbracketed <- stringi::stri_replace_all_regex(
+    string,
+    paste0("\\", bracket, "[^", closing_bracket, "]*\\", closing_bracket),
+    ""
+  )
+
+  return(unbracketed)
 }